requirements.txt
# ======================================================================================
# R-DeepONet Data Factory Requirements (PYAT Version)
# ======================================================================================

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Deep learning frameworks  
torch>=1.12.0
torchvision>=0.13.0

# Data storage and I/O
h5py>=3.6.0
netCDF4>=1.5.8
xarray>=0.20.0

# Visualization and plotting (required for PYAT)
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0
click>=8.0.0

# Optional: Ray distributed computing
ray[default]>=2.0.0

# Optional: Hyperparameter optimization
optuna>=3.0.0

# Experiment tracking (optional)
mlflow>=2.12.0

# Development and testing (optional)
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0

# ======================================================================================
# PYAT Library Notes:
# - PYAT (Python tools for the Acoustic Toolbox) is included as local files in ./PYAT-main/
# - No pip installation required - imported directly via sys.path.append('./PYAT-main')
# - PYAT dependencies: numpy, matplotlib (already included above)
# - External acoustic executables required: bellhopcuda.exe, kraken.exe, scooter.exe
# ======================================================================================